Abstract task storage to Storage class
